requirements.txt
numpy
Pillow
watchdog
